fix: Init user store across all routes
